Fix and clean debouncer
